Update middleware and condition factory classes
In this commit, the classes for middleware and condition factories have been updated. Specifically, the creation process logic has been modified so they can now also handle cases where the name of the class exists without being registered in the container. Additionally, MiddlewareFactory is now used in Stack to create instances, and the user type hint in HasCap's constructor is removed to make it more flexible.